requirements.txt
pyzmq==26.2.0
git+https://github.com/Desklop/StressRNN
